Use runc kill all for process termination
Before this change we would rely on containerd to call delete that also
had logic to kill all processes in a container via cgroups for
containers that run in another pid namespace.  This changes allows us to
decouple the kil all logic and delete so that we don't have to have
containerd calling delete and it can all be handled in the shim again.
This will hopefully prevent races where higher level logic was trying to
remove a containers fs before the shim actually exits.  Now the exit
fifo will not be closed until the shim actually exits.

Signed-off-by: Michael Crosby <crosbymichael@gmail.com>