runtime: allocate cmdDoneCh on loadProcess
In short: loadProcess() does not allocate p.cmdDoneCh so this is _nil_ at
the time when handleSigkilledShim() attempts to read from p.cmdDoneCh.

Long version:

The flow of execution during startup of containerd looks like this.

  daemon (containerd/main.go)
    supervisor.New (supervisor/supervisor.go)
      restore (supervisor/supervisor.go)
        runtime.Load (runtime/container.go)
          loadProcess (runtime/process.go)
            //
            // Note that loadProcess does not allocate cmdDoneCh here.
            // The initial value of cmdDoneCh is nil.
            //
            p := &process{
                    root:      root,
                    id:        id,
                    container: c,
                    spec:      s.ProcessSpec,
                    stdio: Stdio{
                            Stdin:  s.Stdin,
                            Stdout: s.Stdout,
                            Stderr: s.Stderr,
                    },
                    state: Stopped,
            }

            p.ExitStatus (runtime/process.go)
              p.handleSigkilledShim (runtime/process.go)
              //
              // This attempt to read from cmdDoneCh blocks indefinitely
              // because during startup of containerd, this channel does
              // not get allocated and there is no goroutine that would
              // send something to the channel.
              //
              <-p.cmdDoneCh

Commit 9bc2af2e42da82e298f379f8c67385692239b21c introduces this change:

-       <-p.cmdDoneCh
+       select {
+       case <-p.cmdDoneCh:
+       case <-time.After(2 * time.Minute):
+               return rst, fmt.Errorf("could not get the shim ProcessState within two minutes")
+       }

The above change kind of mitigates the problem because the function will
no longer block indefinitely when it attempts to read from a nil channel.
However, this does not address the problem properly because it does not
avoid an attempt to read from a channel that will not succeed anyway.
The startup of containerd will be delayed unnecessarily by at least two
minutes (if the same thing happens for N processes, the delay will even
be N * two minutes). I believe we need a better solution for this.

Signed-off-by: Antonio Murdaca <runcom@linux.com>