Close dead background tasks
If a task is both a background task and has no done channel, assume it
it a dead task and ignore it.  Without this fix a healthcheck with an
unendeding cat command could cause the task queue to eventually fill and
only a reboot would resolve it.

Addresses: https://bugzilla.redhat.com/show_bug.cgi?id=1636244

Signed-off-by: TomSweeneyRedHat <tsweeney@redhat.com>