fix error handling in restore() function
If runtime.Load() returns an error, the restore() function removes
/run/docker/libcontainerd/containerd/CONTAINERID recursively. However,
if the error is not related to the container's init process, then this
function should only remove the CONTAINERID/PROCESSID subdirectory.

This patch adds another return value to runtime.Load() that enables the
restore() function to distinguish which directory should be removed. If:

- return value == "init"
  remove /run/docker/libcontainerd/containerd/CONTAINERID

- return value == PROCESSID of exec'ed process
  remove /run/docker/libcontainerd/containerd/CONTAINERID/PROCESSID

Signed-off-by: Ulrich Obergfell <uobergfe@redhat.com>