fix deadlock in restore() function
When containerd starts up, the restore() function walks through the directory
hierarchy under /run/docker/libcontainerd/containerd and imports the state of
processes from files in /run/docker/libcontainerd/containerd/CONTAINERID and
in /run/docker/libcontainerd/containerd/CONTAINERID/PROCESSID. The restore()
function adds an ExitTask entry to the s.tasks queue for each process that is
no longer in state 'running'. The size of the s.tasks queue is hard-coded and
limited to 2048 (defaultBufferSize). If more than 2048 ExitTask entries need
to be added to the queue, the restore() function gets blocked (queue is full).
If this happens, containerd is in a kind of deadlock situation because the
handleTask() function (which would drain the ExitTask entries from the queue)
has not been started in a separate goroutine yet, and the main goroutine is
blocked in the restore() function (unable to start the handleTask() function).

This patch introduces the dynamically allocated restoreExitTasks slice which
the restore() function uses to store the ExitTask entries separately instead
of adding them to the s.tasks queue. The task handler goroutine subsequently
drains all entries from restoreExitTasks frist before it enters the loop that
handles entries from the s.tasks queue.

Signed-off-by: Ulrich Obergfell <uobergfe@redhat.com>